requirements.txt
openai
fastapi
uvicorn
requests
python-dotenv
python-telegram-bot